basico funcionando despues  del refactoring
