+++ application/extensions.py
@@ -1,0 +1,7 @@
+from flask_sqlalchemy import SQLAlchemy
+from flask_bcrypt import Bcrypt
+from flask_login import LoginManager
+
+db = SQLAlchemy()
+bcrypt = Bcrypt()
+login_manager = LoginManager()
