+++ application/routes.py
@@ -1,88 +1,63 @@
-from application import app, db
-from flask import render_template, redirect, url_for, flash, request
+from application import app
+from flask import render_template, redirect, url_for, flash, request, session
 from flask_login import login_user, logout_user, login_required, current_user
-from application.models import *
-from application.forms import RegisterForm, LoginForm, PurchaseItemForm, SellItemForm
-from application.variables import (
+from .models import User, Role, Subscription, Vehicle, Contractor, PropertyTitle, PropertyDocument, Document
+from .forms import LoginForm, RegisterForm
+from .variables import (
     questions_seller, questions_seller_a, questions_seller_b, questions_seller_c,
     questions_buyer, questions_car, questions_transaction
 )
-from application.transform_answers import transform_document
-from application.seller_answers import get_seller_answers
-from application.buyer_answers import get_buyer_answers
-
+from .transform_answers import transform_document
+from .create_pdf import create_pdf
+from .extensions import db
 
 @app.route('/')
 @app.route('/home')
 def home_page():
+    # Aquí puedes agregar lógica para mostrar información relevante en el home
     return render_template('home.html')
-
-@app.route('/application', methods=['GET', 'POST'])
-@login_required
-def application_page():
-    purchase_form = PurchaseItemForm()
-    selling_form = SellItemForm()
-    if request.method == "POST":
-        #Purchase Item Logic
-        purchased_item = request.form.get('purchased_item')
-        p_item_object = Item.query.filter_by(name=purchased_item).first()
-        if p_item_object:
-            if current_user.can_purchase(p_item_object):
-                p_item_object.buy(current_user)
-                flash(f"Congratulations! You purchased {p_item_object.name} for {p_item_object.price}$", category='success')
-            else:
-                flash(f"Unfortunately, you don't have enough money to purchase {p_item_object.name}!", category='danger')
-        #Sell Item Logic
-        sold_item = request.form.get('sold_item')
-        s_item_object = Item.query.filter_by(name=sold_item).first()
-        if s_item_object:
-            if current_user.can_sell(s_item_object):
-                s_item_object.sell(current_user)
-                flash(f"Congratulations! You sold {s_item_object.name} back to application!", category='success')
-            else:
-                flash(f"Something went wrong with selling {s_item_object.name}", category='danger')
-
-
-        return redirect(url_for('application_page'))
-
-    if request.method == "GET":
-        items = Item.query.filter_by(owner=None)
-        owned_items = Item.query.filter_by(owner=current_user.id)
-        return render_template('home.html', items=items, purchase_form=purchase_form, owned_items=owned_items, selling_form=selling_form)
-
-@app.route('/register', methods=['GET', 'POST'])
-def register_page():
-    form = RegisterForm()
-    if form.validate_on_submit():
-        user_to_create = User(username=form.username.data,
-                              email_address=form.email_address.data,
-                              password=form.password1.data)
-        db.session.add(user_to_create)
-        db.session.commit()
-        login_user(user_to_create)
-        flash(f"Account created successfully! You are now logged in as {user_to_create.username}", category='success')
-        return redirect(url_for('application_page'))
-    if form.errors != {}: #If there are not errors from the validations
-        for err_msg in form.errors.values():
-            flash(f'There was an error with creating a user: {err_msg}', category='danger')
-
-    return render_template('register.html', form=form)
 
 @app.route('/login', methods=['GET', 'POST'])
 def login_page():
     form = LoginForm()
     if form.validate_on_submit():
-        attempted_user = User.query.filter_by(username=form.username.data).first()
-        if attempted_user and attempted_user.check_password_correction(
-                attempted_password=form.password.data
-        ):
-            login_user(attempted_user)
-            flash(f'Success! You are logged in as: {attempted_user.username}', category='success')
-            return redirect(url_for('application_page'))
-        else:
-            flash('Username and password are not match! Please try again', category='danger')
-
+        # Aquí va la lógica de inicio de sesión
+        pass
     return render_template('login.html', form=form)
+
+@app.route('/register', methods=['GET', 'POST'])
+def register_page():
+    form = RegisterForm()
+    if form.validate_on_submit():
+        # Aquí va la lógica de registro
+        pass
+    return render_template('register.html', form=form)
+
+@app.route('/questions/<int:question_id>', methods=['GET', 'POST'])
+def question(question_id):
+    all_questions = questions_seller + questions_seller_a + questions_seller_b + questions_seller_c + questions_buyer + questions_car + questions_transaction
+    if question_id < len(all_questions):
+        current_question = all_questions[question_id]
+        if request.method == 'POST':
+            answer = request.form.get('answer')
+            session['responses'][current_question] = answer
+            session.modified = True
+            next_question_id = question_id + 1
+            if next_question_id < len(all_questions):
+                return redirect(url_for('question', question_id=next_question_id))
+            else:
+                return redirect(url_for('generate_document'))
+        return render_template('questions.html', question=current_question, question_id=question_id)
+    else:
+        # Lógica cuando todas las preguntas hayan sido respondidas
+        return redirect(url_for('generate_document'))
+
+@app.route('/generate_document')
+def generate_document():
+    documento_final = transform_document(session['responses'], documento)
+    output_pdf = "output.pdf"
+    create_pdf(documento_final, output_pdf)
+    return render_template('success.html', pdf_file=output_pdf)
 
 @app.route('/logout')
 def logout_page():
@@ -90,19 +65,14 @@ def logout_page():
     flash("You have been logged out!", category='info')
     return redirect(url_for("home_page"))
 
-@app.route('/questions/<int:question_id>', methods=['GET', 'POST'])
-def question(question_id):
-    all_questions = questions_seller + questions_seller_a + questions_seller_b + questions_seller_c + questions_buyer + questions_car + questions_transaction
-    if question_id < len(all_questions):
-        current_question = all_questions[question_id]
-        # Your logic to display the question and handle the answer
-        return render_template('questions.html', question=current_question, question_id=question_id)
-    else:
-        # Logic for when all questions have been answered
-        return redirect(url_for('some_final_endpoint'))
-    
 @app.route('/create_task', methods=['POST'])
 def create_task():
-    # Implement the logic to handle task creation
+    # Lógica para crear una nueva tarea
     # ...
     return redirect(url_for('home_page'))
+
+
+# Añade cualquier otra ruta que necesites aquí
+
+# No es necesario que incluyas if __name__ == '__main__': aquí,
+# ya que este archivo se importará en __init__.py y no se ejecutará directamente.
